rag/requirements.txt
langchain>=0.0.200
langchain_huggingface>=0.0.5
langchain_ollama>=0.0.5
langchain_community>=0.0.10
beautifulsoup4>=4.12.0
markdownify>=0.11.6
pyyaml>=6.0
faiss-gpu>=1.7.4  # Using GPU version for faster processing